separated postgres and redis repositories
