+++ backend/internal/quickflow.go
@@ -1,85 +1,85 @@
 package internal
 
 import (
-    "fmt"
-    "net/http"
+	"fmt"
+	"net/http"
 
-    "github.com/gorilla/mux"
+	"github.com/gorilla/mux"
 
-    "quickflow/config"
-    qfhttp "quickflow/internal/delivery/http"
-    "quickflow/internal/delivery/http/middleware"
-    "quickflow/internal/repository/postgres_redis"
-    "quickflow/internal/usecase"
+	"quickflow/config"
+	qfhttp "quickflow/internal/delivery/http"
+	"quickflow/internal/delivery/http/middleware"
+	"quickflow/internal/repository/postgres"
+	"quickflow/internal/usecase"
 )
 
 func Run(cfg *config.Config, corsCfg *config.CORSConfig) error {
-    if cfg == nil {
-        return fmt.Errorf("config is nil")
-    }
+	if cfg == nil {
+		return fmt.Errorf("config is nil")
+	}
 
-    //newRepo := repository.NewInMemory()
-    newUserRepo := postgres_redis.NewPostgresUserRepository()
-    newPostRepo := postgres_redis.NewPostgresPostRepository()
-    newSessionRepo := postgres_redis.NewRedisSessionRepository()
-    newAuthService := usecase.NewAuthService(newUserRepo, newSessionRepo)
-    newPostService := usecase.NewPostService(newPostRepo)
-    newAuthHandler := qfhttp.NewAuthHandler(newAuthService)
-    newPostHandler := qfhttp.NewFeedHandler(newPostService, newAuthService)
+	//newRepo := repository.NewInMemory()
+	newUserRepo := postgres.NewPostgresUserRepository()
+	newPostRepo := postgres.NewPostgresPostRepository()
+	newSessionRepo := postgres.NewRedisSessionRepository()
+	newAuthService := usecase.NewAuthService(newUserRepo, newSessionRepo)
+	newPostService := usecase.NewPostService(newPostRepo)
+	newAuthHandler := qfhttp.NewAuthHandler(newAuthService)
+	newPostHandler := qfhttp.NewFeedHandler(newPostService, newAuthService)
 
-    // routing
-    r := mux.NewRouter()
-    r.Use(middleware.CORSMiddleware(*corsCfg))
-    r.MethodNotAllowedHandler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-        http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
-    })
+	// routing
+	r := mux.NewRouter()
+	r.Use(middleware.CORSMiddleware(*corsCfg))
+	r.MethodNotAllowedHandler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
+	})
 
-    r.PathPrefix("/api/").HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-        if r.Method == http.MethodOptions {
-            w.WriteHeader(http.StatusNoContent)
-            return
-        }
-    }).Methods(http.MethodOptions)
+	r.PathPrefix("/api/").HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if r.Method == http.MethodOptions {
+			w.WriteHeader(http.StatusNoContent)
+			return
+		}
+	}).Methods(http.MethodOptions)
 
-    r.HandleFunc("/hello", newAuthHandler.Greet).Methods(http.MethodGet)
+	r.HandleFunc("/hello", newAuthHandler.Greet).Methods(http.MethodGet)
 
-    apiRouter := r.PathPrefix("/").Subrouter()
-    // validating that the content type is application/json for every route but /hello
-    apiRouter.Use(middleware.ContentTypeMiddleware("application/json"))
+	apiRouter := r.PathPrefix("/").Subrouter()
+	// validating that the content type is application/json for every route but /hello
+	apiRouter.Use(middleware.ContentTypeMiddleware("application/json"))
 
-    apiRouter.HandleFunc("/signup", newAuthHandler.SignUp).Methods(http.MethodPost)
-    apiRouter.HandleFunc("/login", newAuthHandler.Login).Methods(http.MethodPost)
-    apiRouter.HandleFunc("/logout", newAuthHandler.Logout).Methods(http.MethodPost)
+	apiRouter.HandleFunc("/signup", newAuthHandler.SignUp).Methods(http.MethodPost)
+	apiRouter.HandleFunc("/login", newAuthHandler.Login).Methods(http.MethodPost)
+	apiRouter.HandleFunc("/logout", newAuthHandler.Logout).Methods(http.MethodPost)
 
-    // Subrouter for protected routes
-    protected := apiRouter.PathPrefix("/").Subrouter()
-    protected.Use(middleware.SessionMiddleware(newAuthService))
+	// Subrouter for protected routes
+	protected := apiRouter.PathPrefix("/").Subrouter()
+	protected.Use(middleware.SessionMiddleware(newAuthService))
 
-    staticHandler := http.StripPrefix("/photos/", http.FileServer(http.Dir("./static/photos")))
+	staticHandler := http.StripPrefix("/photos/", http.FileServer(http.Dir("./static/photos")))
 
-    r.PathPrefix("/photos/").Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-        if r.URL.Path == "/photos/" { // Блокируем просмотр списка файлов
-            http.Error(w, "403 Forbidden", http.StatusForbidden)
-            return
-        }
-        staticHandler.ServeHTTP(w, r)
-    }))
+	r.PathPrefix("/photos/").Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if r.URL.Path == "/photos/" { // Блокируем просмотр списка файлов
+			http.Error(w, "403 Forbidden", http.StatusForbidden)
+			return
+		}
+		staticHandler.ServeHTTP(w, r)
+	}))
 
-    protected.HandleFunc("/feed", newPostHandler.GetFeed).Methods(http.MethodPost)
-    protected.HandleFunc("/post", newPostHandler.AddPost).Methods(http.MethodPost)
+	protected.HandleFunc("/feed", newPostHandler.GetFeed).Methods(http.MethodPost)
+	protected.HandleFunc("/post", newPostHandler.AddPost).Methods(http.MethodPost)
 
-    server := http.Server{
-        Addr:         cfg.Addr,
-        Handler:      r,
-        ReadTimeout:  cfg.ReadTimeout,
-        WriteTimeout: cfg.WriteTimeout,
-    }
+	server := http.Server{
+		Addr:         cfg.Addr,
+		Handler:      r,
+		ReadTimeout:  cfg.ReadTimeout,
+		WriteTimeout: cfg.WriteTimeout,
+	}
 
-    fmt.Printf("starting server at %s\n", cfg.Addr)
-    err := server.ListenAndServe()
-    if err != nil {
-        return fmt.Errorf("internal.Run: %w", err)
-    }
+	fmt.Printf("starting server at %s\n", cfg.Addr)
+	err := server.ListenAndServe()
+	if err != nil {
+		return fmt.Errorf("internal.Run: %w", err)
+	}
 
-    return nil
+	return nil
 }
