+++ backend/internal/repository/postgres/user.go
@@ -1,4 +1,4 @@
-package postgres_redis
+package postgres
 
 import (
 	"context"
@@ -10,7 +10,7 @@ import (
 
 	"quickflow/config"
 	"quickflow/internal/models"
-	pgmodels "quickflow/internal/repository/postgres_redis/postgres-models"
+	pgmodels "quickflow/internal/repository/postgres/postgres-models"
 	"quickflow/utils"
 )
